requirements.txt
langgraph>=0.0.26
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
pydantic>=2.0.0
openai>=1.0.0
anthropic>=0.18.0
openrouter>=0.1.0
requests>=2.31.0
python-dotenv>=1.0.0
pyyaml>=6.0
